scripts/requirements.txt
# Core dependencies
requests>=2.28.0
pyyaml>=6.0
urllib3>=2.0

# For better HTTP handling
httpx>=0.24.0

# For async operations (optional, for future expansion)
aiohttp>=3.8.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
responses>=0.23.0  # For mocking HTTP requests
